+++ src/middleware/global-error-handler.ts
@@ -1,5 +1,5 @@
-import { Response } from '../response'
-import { Middleware } from './index'
+import type { Response } from '../response'
+import type { Middleware } from './index'
 import { configs } from '../index'
 
 export type ErrorHandlerMiddlewareCallback = (response: Response) => boolean