+++ CHANGELOG.md
@@ -6,8 +6,7 @@ Refactored:
   - `mappersmith`: Migrated `middleware` to typescript #306
 
 Added:
-  - `mappersmith`: Added service workers support with default gateway `fetch`
-  - `mappersmith/test`: Added tests for service workers
+  - `mappersmith`: Added service workers support with default gateway `fetch` #311
 
 ## 2.38.1